Refactor ExplainRegistry to only be subscribed once used
Realistically, it's only ever going to be invoked in development,
hence it is wasteful to have a notification subscriber that is
a noop.

Instead we can delay the subscription to when the explain collection
is enabled for the first time, meaning except for a few rare cases
this code won't even be loaded in production environments.
