+++ activerecord/test/cases/explain_subscriber_test.rb
@@ -1,16 +1,15 @@
 # frozen_string_literal: true
 
 require "cases/helper"
-require "active_record/explain_subscriber"
 require "active_record/explain_registry"
 
 if ActiveRecord::Base.lease_connection.supports_explain?
   class ExplainSubscriberTest < ActiveRecord::TestCase
-    SUBSCRIBER = ActiveRecord::ExplainSubscriber.new
+    SUBSCRIBER = ActiveRecord::ExplainRegistry::Subscriber.new
 
     def setup
       ActiveRecord::ExplainRegistry.reset
-      ActiveRecord::ExplainRegistry.collect = true
+      ActiveRecord::ExplainRegistry.start
     end
 
     def test_collects_nothing_if_the_payload_has_an_exception
@@ -19,7 +18,7 @@ if ActiveRecord::Base.lease_connection.supports_explain?
     end
 
     def test_collects_nothing_for_ignored_payloads
-      ActiveRecord::ExplainSubscriber::IGNORED_PAYLOADS.each do |ip|
+      ActiveRecord::ExplainRegistry::Subscriber::IGNORED_PAYLOADS.each do |ip|
         SUBSCRIBER.finish(nil, nil, name: ip)
       end
       assert_empty queries
